Modify setup script to take into account if mysql data is already present from a previous setup

+++ setup.sh
@@ -1,5 +1,9 @@
 #! /usr/bin/env bash
 
+get_env_variable(){
+    echo $(sed -n -e "/${1}/ s/.*\= *//p" .env)
+}
+
 replace_env_variable(){
     sed -i "s~^${1}=.*$~${1}=${2}~" .env
 }
@@ -7,14 +11,39 @@ replace_env_variable(){
 if [ ! -f ./.env ] || [ $1 == "fresh" ]
 then
 
-    # Copy example env file
-    cp ./.env.example ./.env
+    if [ ! -f ./.env ]; then
+        cp ./.env.example ./.env
+    fi
+
+    if [ -f ./docker-compose.override.yml ]
+    then
+        rm ./docker-compose.override.yml
+    fi
 
     # Get the host path for the app
     read -p "Specify the host path to your app: " host_path_app
     host_path_app=$(dirname $(readlink -e ${host_path_app}))/$(basename ${host_path_app})
     echo "Your app directory is: ${host_path_app}"
     replace_env_variable "HOST_PATH_TO_APP" $host_path_app
+
+    # Let user specify where this is being ran
+
+    echo "What sort of environment are running this in?"
+    select environment in "production" "staging" "testing" "development";
+    do
+        case $environment in
+            "production"|"staging"|"testing"|"development")
+            replace_env_variable "APP_ENV" $environment .env;
+            replace_env_variable "NODE_ENV" $environment .env;
+            break;;
+            * ) echo "Invalid selection. EXITING"; rm ./.env; exit;;
+        esac
+    done
+    if [ $environment != "production" ]
+        then
+            cp ./docker-compose.dev.yml ./docker-compose.override.yml
+    fi
+
 
     # Get the PHP version the user wants
     echo "Select a PHP version"
@@ -32,43 +61,42 @@ then
 
      # Get the PHP version the user wants
     echo "Select a MySQL/Maria DB version"
-    select php_version in "Maria DB 10.1" "MySQL 5.7";
+    select mysql_version in "Maria DB 10.1" "MySQL 5.7";
     do
-        case $php_version in
-            "Maria DB 10.1") break;;
+        mysql_version_number=""
+        case $mysql_version in
+            "Maria DB 10.1")break;;
             "MySQL 5.7")
-                replace_env_variable "MYSQL_VERSION" "57" .env;
+                mysql_version_number="57";
                 break;;
             * ) echo "Invalid selection. EXITING"; rm ./.env; exit;;
         esac
     done
 
-    # Get the name the user wants for the MySQL DB name
-    read -p "What should your MySQL DB be called? " mysql_db_name
-    replace_env_variable "MYSQL_DATABASE" $mysql_db_name
+    if [ ! $mysql_version_number == "" ]; then
+        replace_env_variable "MYSQL_VERSION" $mysql_version_number .env;
+    fi
 
-    # Get the name the user wants for the MySQL DB name
-    read -p "What should your MySQL user be called? " mysql_user
-    replace_env_variable "MYSQL_USER" $mysql_user
+    if [ !  -d "${host_path_app}/data/mysql${mysql_version_number}"   ] ; then
+        # Get the name the user wants for the MySQL DB name
+        read -p "What should your MySQL DB be called? " mysql_db_name
+        replace_env_variable "MYSQL_DATABASE" $mysql_db_name
 
-    # Get the name the user wants for the MySQL DB name
-    read -p "What should the password be for the MySQL user '${mysql_user}'? " mysql_password
-    replace_env_variable "MYSQL_PASSWORD" $mysql_password
+        # Get the name the user wants for the MySQL DB name
+        read -p "What should your MySQL user be called? " mysql_user
+        replace_env_variable "MYSQL_USER" $mysql_user
 
-    # Get the name the user wants for the MySQL DB name
-    read -p "What should the password be for the MySQL root user? " mysql_root_password
-    replace_env_variable "MYSQL_ROOT_PASSWORD" $mysql_root_password
+        # Get the name the user wants for the MySQL DB name
+        read -p "What should the password be for the MySQL user '${mysql_user}'? " mysql_password
+        replace_env_variable "MYSQL_PASSWORD" $mysql_password
 
-    echo "Select a node environment?"
-    select node_environment in "production" "staging" "testing" "development";
-    do
-        case $node_environment in
-            "production"|"staging"|"testing"|"development") replace_env_variable "NODE_ENV" $node_environment .env; break;;
-            * ) echo "Invalid selection. EXITING"; rm ./.env; exit;;
-        esac
-    done
+        # Get the name the user wants for the MySQL DB name
+        read -p "What should the password be for the MySQL root user? " mysql_root_password
+        replace_env_variable "MYSQL_ROOT_PASSWORD" $mysql_root_password
+    fi
 fi
+
 
 echo "Building images ... "
 docker-compose build
-echo "Images are all built. elephant-whale is available from you app directory now"
+echo "Images are all built. Now run docker-compose up to bring up your containers"
